+++ test/uploaded_file_test.rb
@@ -70,14 +70,6 @@ describe Shrine::UploadedFile do
     assert_instance_of @uploader.class, uploaded_file.uploader
   end
 
-  it "doesn't attempt to delete itself twice" do
-    uploaded_file = @uploader.upload(fakeio("image"))
-    uploaded_file.storage.expects(:delete).once
-
-    uploaded_file.delete
-    uploaded_file.delete
-  end
-
   it "can be replaced with another file" do
     uploaded_file = @uploader.upload(fakeio("image"), location: "key")